+++ web-app/src/components/Observations/Observations.js
@@ -1,0 +1,58 @@
+import React, { Component } from 'react';
+import _ from 'lodash';
+
+import Title from '../Title';
+import { Button } from '../Form';
+import Icon from '../Icon';
+
+import api from '../../utils/api';
+
+import './Observations.css';
+
+class Observations extends Component {
+  constructor(...props) {
+    super(...props);
+
+    this.state = {
+      observations: [],
+    };
+  }
+
+  vote(value) {
+    const observation = _.last(this.state.observations);
+
+    this.setState({
+      observations: [..._.dropRight(this.state.observations)],
+    });
+
+    api.post('/votes', {
+      body: {
+        observation_id: observation.id,
+        value,
+      },
+    });
+  }
+
+  render() {
+    const observation = _.last(this.state.observations);
+
+    return (
+      <div className="Observations">
+        <Title name="Observations" />
+        {observation && (
+          <div>
+            <div className="Observations__Picture">
+              <img src={`${process.env.REACT_APP_API_URL}/observations/${observation.id}/picture`} alt="Observation" />
+            </div>
+            <div className="Observations__Buttons">
+              <Button onClick={() => this.vote(1)} circle><Icon name="thumbs-up" /></Button>
+              <Button onClick={() => this.vote(0)} >SKIP</Button>
+              <Button onClick={() => this.vote(-1)} circle><Icon name="thumbs-down" /></Button>
+            </div>
+          </div>
+        )}
+      </div>
+    );
+  }
+}
+export default Observations;
